requirements.txt
langchain==0.3.25
langchain-openai==0.3.16
langchain-community==0.3.24
langchain-chroma==0.2.4
chromadb>=1.0.9
streamlit==1.45.0
openai==1.75.0
python-dotenv==1.1.0
tiktoken==0.9.0
psycopg[binary]
passlib[bcrypt]
email-validator
 